build: Adjust Makefile and spec template
Clean up the interaction between the Makefile and the rpkg spec template. The RPM %{make_install} macro expands to define DESTDIR automatically, so we can use that in the Makefile to stage the installation into the correct location. Fix up some prefix and libdir definitions. Remove the seemingly unncessary .spec file too. The rpkg template should be used by COPR.