fix: ensure _run_data is called on a thread
_run_data is not async-aware, so declaring it as an async function and
running it in an event loop results in synchronous execution of the
entire function. This means the Send function only returns once the
entire _run_data function has completed. Running it on the event loop
using run_in_executor runs the function call in a thread executor pool,
ensuring that it doesn't block the calling thread. This is a fairly
naive solution; a better implementation might be able to pass a future
into _run_data, immediately fulfill that future, let the calling
function (Send) return the RPC receipt, and continue running _run_data.