refactor: Use maps to marshal events
ansible-runner events contain a log of JSON that is immaterial to
the worker, but matters to console-dot applications. Using a struct
to marshal the events into discarded all the fields that were not
member fields of the Event struct. Instead, marshal the entire JSON
content into a map. This pulls in all the optional and immaterial
fields.