refactor: Change HTTP post strategy
Ingress expects a multipart HTTP request body instead of plain JSON.
Additionally, the events are cached until the final message and then
posted. It appears that platbook-dispatcher does not expect messages
to arrive one-by-one, but instead batched up together. This change
caches each JSON event as it happens, but only POSTs the entire
batch when the job completes.