refactor: rewrite ansible package
The internal/ansible package now exports a Runner type that holds data
relevant to a specific run. This type will invoke ansible-runner as a
forked background process and read events from the artifacts directory
instead of trying to hold the child process in the foreground. This
reimplementation more clearly describes what the run is doing by
breaking up the responsibility into methods on a data type.

The logic of Run is roughly:
- Create the files that will be watched: job_events and status
- Start watching job_events and status for events
- Send the "executor_on_start" event
- Start ansible-runner

The main logic exists in the handler functions called by the watch
method.

handleJobEvent marshals the event into a data structure, modifies it for
compatibility with playbook-dispatcher and sends it to the Events
channel.

handleStatusFIle reads the value of the status file when it is written.
If the status is "failed", a "executor_on_failed" event is sent to the
Events channel. This handler then sends signals to the jobEvents
channel to exit, and closes the Events channel (signalling to callers
that the job is complete).